game over

+++ database/migrations/2021_03_05_042814_create_users_table.php
@@ -19,7 +19,7 @@ class CreateUsersTable extends Migration
             $table->string('id_card')->unique();
             $table->date('birthday')->format('Y/m/d');
             $table->string('address');
-            $table->bigInteger('amount');
+            $table->decimal('amount',11);
             $table->string('email')->unique();
             $table->string('phone');
             $table->string('password');
